Move logging utilities to own module
This will allow doing more fancy logging-related things in a central
location.
